Adjustment to Relationships and minor edits
This adds the phrase $table='users' to the student and  instuctor model to ensure that the right is used with with the models. The next change was the change in how relationships are handle. The naming conventions were not correct to laravel standards to take advantage of pivot functionality of it, so that laravel does the work of table for you. There is special way to declare the corresponding keys and table so it does. That was completed. The next change was to the show.blade.php file that now includes a list of studnets in the course and instructor names.